Read initial page data from static html —> avoid initial request
Previously even though the page was staticly rendered, on first load the router would load the page content and re-render the page. This commit creates a LRU cache on the API and pre-populates it with the content of the page from the static HTML. It avoids an extra request and load before the interactivity starts working, and it avoids extra requests when the user moves between pages.